SW-6059 Refactor of Workflow to allow for multiple pages with re-usable components - Part 1
