+++ cypress/e2e/client-tasks/pagination.cy.js
@@ -2,7 +2,7 @@ describe("Pagination", () => {
   beforeEach(() => {
     cy.setCookie("Other", "other");
     cy.setCookie("XSRF-TOKEN", "abcde");
-    cy.visit("/supervision/workflow/1");
+    cy.visit("/supervision/workflow/client-tasks");
   });
 
   describe("First page, ellipses and final page", () => {
