+++ cypress/e2e/client-tasks/team_selection.cy.js
@@ -2,7 +2,7 @@ describe("Team Selection", () => {
   beforeEach(() => {
       cy.setCookie("Other", "other");
       cy.setCookie("XSRF-TOKEN", "abcde");
-      cy.visit("/supervision/workflow/1");
+      cy.visit("/supervision/workflow/client-tasks");
   });
 
     it("can be changed to another team", () => {
