+++ cypress/e2e/workflow.cy.js
@@ -2,7 +2,11 @@ describe("Work flow", () => {
   beforeEach(() => {
       cy.setCookie("Other", "other");
       cy.setCookie("XSRF-TOKEN", "abcde");
-      cy.visit("/supervision/workflow/1");
+      cy.visit("/supervision/workflow/");
+  });
+
+  it("homepage redirects to client tasks page", () => {
+    cy.url().should('contain', '/supervision/workflow/client-tasks')
   });
 
   it("shows user that is logged in within banner", () => {
@@ -33,5 +37,5 @@ describe("Work flow", () => {
   
   it("the nav link should contain logout", () => {
     cy.get(".moj-header__navigation-list > :nth-child(3) > a").should("have.attr", "href", "http://localhost:8080/auth/logout")
-  }) 
+  })
 });
