+++ internal/server/wrap_handler.go
@@ -1,0 +1,94 @@
+package server
+
+import (
+	"fmt"
+	"github.com/ministryofjustice/opg-sirius-workflow/internal/sirius"
+	"go.uber.org/zap"
+	"net/http"
+	"time"
+)
+
+type ErrorVars struct {
+	App   WorkflowVars
+	Code  int
+	Error string
+}
+
+type RedirectError string
+
+func (e RedirectError) Error() string {
+	return "redirect to " + string(e)
+}
+
+func (e RedirectError) To() string {
+	return string(e)
+}
+
+type StatusError int
+
+func (e StatusError) Error() string {
+	code := e.Code()
+
+	return fmt.Sprintf("%d %s", code, http.StatusText(code))
+}
+
+func (e StatusError) Code() int {
+	return int(e)
+}
+
+type Handler func(app WorkflowVars, w http.ResponseWriter, r *http.Request) error
+
+func wrapHandler(client Client, logger *zap.Logger, tmplError Template, prefix, siriusURL string, defaultTeamId int) func(next Handler) http.Handler {
+	return func(next Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			start := time.Now()
+
+			var errVars ErrorVars
+
+			vars, err := NewWorkflowVars(client, r, defaultTeamId)
+			if err == nil {
+				errVars.App = *vars
+				err = next(*vars, w, r)
+			}
+
+			sugar := logger.Sugar()
+			sugar.Infow(
+				"Application Request",
+				"method", r.Method,
+				"uri", r.URL.RequestURI(),
+				"duration", time.Since(start),
+			)
+
+			if err != nil {
+				if err == sirius.ErrUnauthorized {
+					http.Redirect(w, r, siriusURL+"/auth", http.StatusFound)
+					return
+				}
+
+				if redirect, ok := err.(RedirectError); ok {
+					http.Redirect(w, r, prefix+redirect.To(), http.StatusFound)
+					return
+				}
+
+				sugar.Errorw("Error handler", err)
+
+				code := http.StatusInternalServerError
+				if status, ok := err.(StatusError); ok {
+					if status.Code() == http.StatusForbidden || status.Code() == http.StatusNotFound {
+						code = status.Code()
+					}
+				}
+
+				w.WriteHeader(code)
+				errVars.Code = code
+				errVars.Error = err.Error()
+				err = tmplError.ExecuteTemplate(w, "page", errVars)
+
+				if err != nil {
+					sugar.Errorw("Failed to render error template", err)
+					http.Error(w, err.Error(), http.StatusInternalServerError)
+				}
+			}
+		})
+	}
+}
