+++ web/template/client-tasks.gotmpl
@@ -1,4 +1,4 @@
-{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.WorkflowVars*/ -}}
+{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.ClientTasksVars*/ -}}
 {{ template "page" . }}
 
 {{ define "title" }} OPG Sirius Workflow {{ end }}
@@ -20,12 +20,12 @@
           </div>
           <div class="govuk-grid-row">
             <div class="govuk-grid-column-three-quarters">
-              {{ template "error-summary" .Errors }}
+              {{ template "error-summary" .App.Errors }}
             </div>
           </div>
           <div class="govuk-grid-row">
             <div id="filters-column" class="govuk-grid-column-one-quarter">
-              {{ template "filters" . }}
+              {{ template "client-tasks-filters" . }}
             </div>
 
             <div class="govuk-grid-column-three-quarters">
