+++ web/template/error.gotmpl
@@ -1,4 +1,4 @@
-{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.errorVars*/ -}}
+{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.ErrorVars*/ -}}
 {{ template "page" . }}
 
 {{ define "title" }}
