+++ web/template/layout/due-date.gotmpl
@@ -1,4 +1,4 @@
-{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.WorkflowVars*/ -}}
+{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.ClientTasksVars*/ -}}
 {{ define "due-date" }}
 <div class="moj-filter__options" data-filter-name="moj-filter-name-due-date">
