+++ web/template/layout/task-type.gotmpl
@@ -1,4 +1,4 @@
-{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.WorkflowVars*/ -}}
+{{- /*gotype: github.com/ministryofjustice/opg-sirius-workflow/internal/server.ClientTasksVars*/ -}}
 {{ define "task-type" }}
 
 <div class="moj-filter__options" data-filter-name="moj-filter-name-tasktype">
